// src/payload-types.ts
export interface Occupancy {
  id: number;
  name?: string | null;
  address?: string | null;
  location?: {
    lat?: number | null;
    lng?: number | null;
  };
  occupancyPk: string;
  occupancyId?: string | null;
  validated:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  markers?: {
    docs?: (number | Marker)[] | null;
    hasNextPage?: boolean | null;
  } | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "markers".
 */
export interface Marker {
  id: number;
  dataset: 'fmp' | 'google';
  location?: {
    lat?: number | null;
    lng?: number | null;
  };
  occupancyPk: string;
  occupancyId?: string | null;
  distance?: number | null;
  occupancy?: (number | Occupancy)[] | null;
  updatedAt: string;
  createdAt: string;
}